v0.9.0 Typescript and webpack
Moved the source files into the "src" directory.

Added typescript and the eslint plugins and configs to support it for
node.

Added webpack to build the files from typescript.

Created a node script to execute webpack and easily configure the
process through a config file and plugins.

Nodemon config removed from package.json and used instead as a webpack
plugin. This is so it will run only when webpack watch is enabled, which
is only in development mode.