v0.12.0 Update DB, .env files, workflows
Removed the production .env file which had the database connection info.
This was a security problem because now it is visibile in the
repository. Now, it is a environment variable in Heroku.

Fixed tsconfig.json by keeping the outDir property.

Updated the workflows and added the auto-merge.yml file to automatically
squash merge dependabot pull requests that pass the build workflow.

Updated the database file to use the environment variable in Heroku.

Updated the development .env file variable for the database connection
to work for both production and development.