v0.18.0 Update courses data
Turned off dependabot by commenting out the code.

Added SQL queries to create the courses table and insert the data.

Created the courses controller for the routes and retrieving data.

Updated the routes for the new courses routes.

Updated the server file so that it will default to use a single thread
unles specifically opting in.